refactor: ResponseErrorHandler 추가 및 baseUrl 수정
