+++ src/main/java/roomescape/config/PaymentConfig.java
@@ -1,8 +1,5 @@
 package roomescape.config;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.Base64;
@@ -12,16 +9,14 @@ import org.springframework.boot.web.client.ClientHttpRequestFactorySettings;
 import org.springframework.boot.web.client.RestClientCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.http.client.JdkClientHttpRequestFactory;
+import org.springframework.web.client.ResponseErrorHandler;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.support.RestClientAdapter;
 import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-import roomescape.dto.response.reservation.TossExceptionResponse;
-import roomescape.exception.PaymentException;
+import roomescape.exception.PaymentExceptionHandler;
 import roomescape.service.PaymentService;
 
 @Configuration
@@ -31,9 +26,24 @@ public class PaymentConfig {
 
     @Bean
     public PaymentService paymentService(RestClient.Builder restClientBuilder) {
-        RestClient restClient = restClientBuilder.build();
+        RestClient restClient = restClientBuilder
+                .build();
         HttpServiceProxyFactory factory = HttpServiceProxyFactory.builderFor(RestClientAdapter.create(restClient)).build();
         return factory.createClient(PaymentService.class);
+    }
+
+    @Bean
+    public RestClientCustomizer restClientCustomizer() {
+        return (restClientBuilder) -> restClientBuilder
+                .requestFactory(clientHttpRequestFactory())
+                .defaultHeader(HttpHeaders.AUTHORIZATION, authorization())
+                .defaultStatusHandler(responseErrorHandler())
+                .baseUrl("https://api.tosspayments.com");
+    }
+
+    private String authorization() {
+        byte[] encodedBytes = Base64.getEncoder().encode((widgetSecretKey + ":").getBytes(StandardCharsets.UTF_8));
+        return "Basic " + new String(encodedBytes);
     }
 
     private ClientHttpRequestFactory clientHttpRequestFactory() {
@@ -43,23 +53,7 @@ public class PaymentConfig {
         return ClientHttpRequestFactories.get(JdkClientHttpRequestFactory.class, settings);
     }
 
-    @Bean
-    public RestClientCustomizer restClientCustomizer() {
-        byte[] encodedBytes = Base64.getEncoder().encode((widgetSecretKey + ":").getBytes(StandardCharsets.UTF_8));
-        String authorization = "Basic " + new String(encodedBytes);
-
-        return (restClientBuilder) -> restClientBuilder
-                .requestFactory(clientHttpRequestFactory())
-                .baseUrl("https://api.tosspayments.com/v1/payments")
-                .defaultHeader("Authorization", authorization)
-                .defaultStatusHandler(HttpStatusCode::isError, ((request, response) -> {
-                    throw new PaymentException((HttpStatus) response.getStatusCode(), getTossExceptionResponse(response));
-                }));
-    }
-
-    private TossExceptionResponse getTossExceptionResponse(ClientHttpResponse response) throws IOException {
-        return new ObjectMapper()
-                .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-                .readValue(response.getBody(), TossExceptionResponse.class);
+    private ResponseErrorHandler responseErrorHandler() {
+        return new PaymentExceptionHandler();
     }
 }
