+++ src/main/java/roomescape/exception/PaymentExceptionHandler.java
@@ -1,0 +1,22 @@
+package roomescape.exception;
+
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.io.IOException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.client.ClientHttpResponse;
+import org.springframework.web.client.DefaultResponseErrorHandler;
+import roomescape.dto.response.reservation.TossExceptionResponse;
+
+public class PaymentExceptionHandler extends DefaultResponseErrorHandler {
+    @Override
+    public void handleError(ClientHttpResponse response) throws IOException {
+        throw new PaymentException((HttpStatus) response.getStatusCode(), getTossExceptionResponse(response));
+    }
+
+    private TossExceptionResponse getTossExceptionResponse(ClientHttpResponse response) throws IOException {
+        return new ObjectMapper()
+                .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+                .readValue(response.getBody(), TossExceptionResponse.class);
+    }
+}
